Variable detection of long double Nan/INF signatures
libast attempts to determine the binary representation of Inf and NaN to
use as a fall-back code path for systems that do not support fpclassify().
That code path is not used on illumos but the libast feature detection does
not get consistent signatures and they change between builds.

To fix this, zero the memory before determining the signature.

Note that the fall-back code path is broken because there are multiple
representations for NaN - the important thing is to check the exponent
and for a non-zero significand. The trailing bits can be random or
left over from interim operations.

For that reason, to ensure we never end up using the fall-back code path
explicitly generate a compile error if we end up there.